+++ R/imports.R
@@ -6,6 +6,7 @@
 #' @import cli
 #' @import purrr
 #' @import stringr
+#' @import lme4
 #' @importFrom dplyr filter lag
 #' @importFrom magrittr %>%
 #' @importFrom stats setNames complete.cases pnorm lm nobs as.formula